requirements.txt
transformers==4.39.2
dataset==2.18.0
tokenizers==0.15.2
torch==2.2.2
unidecode